requirements.txt
streamlit
pandas
numpy
google-api-python-client
google-generativeai
google-genai
sentence-transformers
scikit-learn
umap-learn
jieba
plotly